requirements.txt
streamlit
firebase-admin
setuptools
matplotlib
google-genai